[op][agl][wayland] Take into account wayland output transform
To establish system display resolution, Wayland backend uses the
wl_output mode (to identify current mode). But that is stating the
physical resolution of the display. It can be transformed, so the
display resolution that Chromium should be aware of the rotation
to swap the axis in that case.

The transformation is reported with the wl_output geometry event.

[SPEC-4224] Adapt WAM to new application framework
https://jira.automotivelinux.org/browse/SPEC-4224